refactor(pkg): move ctxcancel from pkg/persistence/ to pkg/ (PR271-FU4)
Closes the FU4 placement evaluation. Original backlog framing ("revisit when
a second ctxcancel-flavored helper appears") was a pseudo-trigger — future
helpers (pgx error classifier, retry classifier) belong elsewhere by their
own dependency profile, so the trigger never fires. Decided to move now.

Why pkg/ (not kernel/ or runtime/):

- pkg/httputil already imports ctxcancel.DetailsKeyReason; pkg/ is forbidden
  from depending on kernel/ or runtime/ (CLAUDE.md layering rule), so any
  upward move would break the existing pkg→pkg dependency chain — would
  force httputil to either hardcode the "reason" key (re-introducing the
  string-duplication smell the constant was extracted to fix) or strip the
  helper API surface.
- ctxcancel is stateless pure functions on standard library + pkg/errcode;
  kernel/ is "Cell/Slice runtime + governance" (architecture core) and
  runtime/ is "stateful runtime facilities (auth verifier / outbox / tracer)"
  — neither matches ctxcancel's shape.
- Unlike kernel/persistence/TxCtxKey (which exists specifically to bridge
  the cells/ ↔ adapters/ no-import rule via a kernel-owned context key),
  ctxcancel has no such layering conflict — every consumer (cells/, pkg/,
  runtime/) is already permitted to import pkg/.

Why drop the persistence/ segment:

- Package doc explicitly scopes the helper to "IO operations (DB scan,
  RPC call, message bus claim)" — the persistence/ segment narrows it
  artificially to DB.
- All other pkg/ siblings are flat capability-typed: aeadutil, ctxkeys,
  errcode, httputil, idutil, query, securecookie, testutil, validation.
  pkg/persistence/ was the lone exception with a single child.
- Empty pkg/persistence/ parent removed.

Mechanical changes:

- git mv pkg/persistence/ctxcancel pkg/ctxcancel
- 8 import path replacements: 4 production sources (configcore × 2 +
  auditcore × 1 + pkg/httputil × 1) + 2 tests (auditcore ctx-cancel +
  tracing_test) + 2 doc ref-comments
- gofmt -w on the 6 touched Go files (alphabetical pkg/ctxcancel vs
  pkg/ctxkeys/errcode/httputil reordering)
- docs/backlog.md FU4 entry rewritten: from "decision: stay" to
  "migrated", with the layering analysis recorded
- docs/plans/202604252100-026-post-v1.0-cleanup-plan.md updated path refs

Validation:
- go build ./... — clean
- go test ./pkg/ctxcancel/... ./pkg/errcode/... ./pkg/httputil/...
  ./runtime/http/middleware/... ./cells/auditcore/...
  ./cells/configcore/... -count=1 — all pass
- golangci-lint run on the same packages — 0 issues
