+++ docs/plans/202604252100-026-post-v1.0-cleanup-plan.md
@@ -299,14 +299,14 @@ PR-A6 typed event payload 落地 + PR-A5c outbox emitter 收口之后的 trailin
 - `PR238-FU3` / `FU4` / `FU6` / `FU7` / `FU8`（configcore 各 ctx-cancel 处错误归类与文档对齐）
 - `STALECIPHER-OPFIELD`（stale cipher 错误 op 字段统一）
 - `FLAGREPO-CTXCANCEL`（flagrepo ctx-cancel 路径统一）
-- `CTXCANCEL-HELPER-PKGLEVEL`（决策抽 `pkg/persistence/ctxcancel.go` 公共 helper）
+- `CTXCANCEL-HELPER-PKGLEVEL`（决策抽 `pkg/ctxcancel.go` 公共 helper）
 - `PR238-FU2`（infra bucket category 漂移识别）+ `FU2-GOVERNANCE-STATIC` OBS-IS-INFRA-BUCKET-GOVERNANCE-STATIC-01（`kernel/governance/rules_obs.go` 扫 Prometheus counter 注册的 label 来源）
 
-**前置决策**：是否抽 `pkg/persistence/ctxcancel.go`（影响所有 cell repo），需 architect 评审。
+**前置决策**：是否抽 `pkg/ctxcancel.go`（影响所有 cell repo），需 architect 评审。
 
 **验收**：
 - configcore 所有 repo 的 ctx-cancel 路径统一 `Category: CategoryInfra` 或 `CategoryAuth` 分类决策
-- `pkg/persistence/ctxcancel` helper 接口稳定，三 cell（accesscore/configcore/auditcore）全部迁移
+- `pkg/ctxcancel` helper 接口稳定，三 cell（accesscore/configcore/auditcore）全部迁移
 - `kernel/governance/rules_obs.go` archtest 扫到 Prometheus counter label 来源不在 errcode bucket 时 PR fail
 - bucket-migration 表写入 PR description
 
@@ -599,7 +599,7 @@ PR-A12b kernel/command 生命周期收口之后的生产治理项；触发条件
 
 - **PR-A47 ERRCODE-5XX-MASK** 需要 PR-A50 / PR-A51 ctx-cancel 收口前完成，避免 503 → 499 重映射时 errcode 字典再次翻新
 - **PR-A49 OUTBOX-ARCHTEST-EXPAND** 的 AST 升级路径参考 PR246-FU1 FMT-19 已落地实现，无前置阻塞
-- **PR-A50 CONFIGCORE-CTX-CANCEL-CONSOLIDATE** 抽 pkg 级 helper 需 architect 决策（`pkg/persistence/ctxcancel.go`）
+- **PR-A50 CONFIGCORE-CTX-CANCEL-CONSOLIDATE** 抽 pkg 级 helper 需 architect 决策（`pkg/ctxcancel.go`）
 - **PR-A53 BOOTSTRAP-LISTENER-SLICE-POLISH** 的 Checker 签名升级需要 `kernel/lifecycle.ManagedResource.Checkers()` 同步升级，影响所有 adapter — 与 PR-A54 同 Sprint 执行
 
 ### 触发条件项的依赖
